The Mutation object for snp_data and snp_anno

The Mutation object is created to apply all functions for mutation analysis

Public fields

snp_list

A numeric field

metadata

A numeric field

modification_info

A non-numeric field

nj_distance

NJ distance Initialize a new MyClass object

Methods

Public methods

SNP$new()

SNP$get_metadata()

SNP$ML_data_dt()

SNP$Evolution_dt()

SNP$filter_snp()

SNP$filter_sample()

SNP$estimate_imp_all()

SNP$estimate_imp()

SNP$MutTree()

SNP$MutNumPlot()

SNP$calculate_NJ()

SNP$clone()

Method new()

Usage

SNP$new(
  snp_anno,
  snp_data,
  metadata = NULL,
  remove_genes_dt = NA,
  annotate_AA = T,
  verbose = F
)

Arguments

snp_anno

Initial value for the field snp_anno

snp_data

Initial value for the field snp_data

metadata

Initial value for the field metadata

remove_genes_dt

Initial value for the field remove_genes_dt

annotate_AA

Initial value for the field remove_genes_dt

verbose

Initial value for the field remove_genes_dt get_metadata

modification_info

Initial value for the field remove_genes_dt

Method get_metadata()

Usage

SNP$get_metadata()

Arguments

Initialize

a new MyClass object

Method ML_data_dt()

Usage

SNP$ML_data_dt(selected_pheno)

Arguments

selected_pheno

Initial value for the field selected_pheno

Returns

NULL Initialize a new MyClass object

Method Evolution_dt()

Usage

SNP$Evolution_dt(x)

Arguments

x

Initial value for the field x Getting SNPs in certain Genes, or certain SNP in all samples

Method filter_snp()

Usage

SNP$filter_snp(gene, gene_type = "INDEX", upstream_length = NULL)

Arguments

gene

values for Index , Locus_tag or Gene

gene_type

values for searching mutations by certain type, should be one of "GENE","LOCUS_TAG","INDEX"

upstream_length

whether to show the upstream mutation of genes

length

Length of upstream and downstream regions of genes that should be included Initialize a new MyClass object

Method filter_sample()

Usage

SNP$filter_sample(sample_name = NULL, mutation_index = NULL, opposite = F)

Arguments

sample_name

sample names to filter

mutation_index

mutation index.

opposite

Used with mutation_index. When opposite is set to TRUE, samples containing mutation_index will be removed; otherwise will be remained. Getting SNPs counts in all samples

Method estimate_imp_all()

Usage

SNP$estimate_imp_all()

Method estimate_imp()

Usage

SNP$estimate_imp(
  gene,
  gene_type = "GENE",
  metadata_col = NULL,
  upstream_length = NULL
)

Arguments

gene

values for Index , Locus_tag or Gene

gene_type

values for searching mutations by certain type, should be one of "GENE","LOCUS_TAG","INDEX"

metadata_col

Default is NULL. If the metadata_col was set, the estimate_imp function will calculate significance difference of mutation counts among different metadata_col levels. Should be colnames in metadata or the mutation INDEX.

upstream_length

whether to show the upstream mutation of genes MutTree plot phylogenetic tree with heatmap of snps and lineage composition

Method MutTree()

Usage

SNP$MutTree(locus_tag, limit = 5, size = 20, metadata_col = NULL)

Arguments

locus_tag

the locus_tag of wanted gene

limit

the minimum frequence a sample need to have,defaults to 5

size

the number of top occured mutation types to plot,defaults to 20

metadata_col

Default is NULL. If the metadata_col was set, the estimate_imp function will calculate significance difference of mutation counts among different metadata_col levels. Should be colnames in metadata or the mutation INDEX. MutNumPlot plot mutation times and modification types in a single gene.

Method MutNumPlot()

Usage

SNP$MutNumPlot(
  locus_tag,
  limit = 1,
  label_num = 10,
  show_synomonous = T,
  Modification = NULL
)

Arguments

locus_tag

the locus_tag of wanted gene

limit

the minimum frequence a sample need to have,defaults to 1

label_num

the number of top frequent mutationsto label, defaults to 10

show_synomonous

Whether to show synomonous mutations.

Modification

Modification types to show. calculate_NJ calculate sample distance for object based on Neighbor-Joining Algorithm.

metadata_col

Default is NULL. If the metadata_col was set, the estimate_imp function

Method calculate_NJ()

Usage

SNP$calculate_NJ(num_thread = 1, snp_range = c(0.05, 0.95))

Arguments

num_thread

number of threads to use

snp_range

SNPs within this frequency range were included in the NJ (Neighbor-Joining) distance calculation. This range was used to accelerate computational speed.

Method clone()

The objects of this class are cloneable with this method.

Usage

SNP$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.